refactor: 알림 관련 로직을 커스텀 훅으로 분리
- notifee 이벤트 핸들링을 useNotifeeEvents 훅으로 분리
- 권한 요청 로직을 useNotificationPermissions 훅으로 분리
- 진동 채널 생성 로직을 useVibrationChannels 훅으로 분리
- App 컴포넌트가 더 간결하고 책임이 분리된 구조로 개선됨
