+++ frontend/App.tsx
@@ -5,10 +5,7 @@
  * @format
  */
 
-import notifee, {
-  AndroidNotificationSetting,
-  EventType,
-} from '@notifee/react-native';
+import notifee from '@notifee/react-native';
 import {
   LinkingOptions,
   NavigationContainer,
@@ -16,10 +13,13 @@ import {
 } from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import React, {useEffect, useRef} from 'react';
-import {Alert, Linking, PermissionsAndroid, Platform} from 'react-native';
+import {Linking} from 'react-native';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
 import {enableScreens} from 'react-native-screens';
 
+import useNotifeeEvents from './src/hooks/useNotifeeEvents';
+import useNotificationPermissions from './src/hooks/useNotificationPermissions';
+import useVibrationChannels from './src/hooks/useVibrationChannels';
 import AuthStack from './src/navigation/AuthStack';
 import HomeTabs from './src/navigation/HomeTabs';
 import AppLockScreen from './src/screens/AppLockScreen';
@@ -36,7 +36,6 @@ import SetAppLockPasswordScreen from './src/screens/main/settings/SetAppLockPass
 import SplashScreen from './src/screens/SplashScreen';
 import {useAppLockStore} from './src/store/appLockStore';
 import {useAuthStore} from './src/store/authStore';
-import {createVibrationChannels} from './src/utils/alarmManager';
 
 export type RootStackParamList = {
   Auth: undefined;
@@ -110,6 +109,10 @@ const App = () => {
 
   const navigationRef = useRef<NavigationContainerRef<any>>(null);
 
+  useNotificationPermissions();
+  useNotifeeEvents(navigationRef);
+  useVibrationChannels();
+
   useEffect(() => {
     checkAuth();
   }, [checkAuth]);
@@ -117,100 +120,6 @@ const App = () => {
   useEffect(() => {
     checkAppLock();
   }, [isLoggedIn, checkAppLock]);
-
-  useEffect(() => {
-    // Background 이벤트 처리
-    notifee.onBackgroundEvent(async ({type, detail}) => {
-      if (
-        type === EventType.PRESS ||
-        (type === EventType.ACTION_PRESS && detail.pressAction?.id === 'accept')
-      ) {
-        console.log('Background 이벤트 처리', detail.notification?.data);
-        navigationRef.current?.navigate('IncomingCall');
-      }
-    });
-
-    // Foreground 이벤트 처리
-    const unsubscribeFg = notifee.onForegroundEvent(async ({type, detail}) => {
-      if (type === EventType.DELIVERED || type === EventType.PRESS) {
-        if (
-          detail.notification?.data?.link &&
-          typeof detail.notification?.data?.link === 'string' &&
-          detail.notification?.data?.link.includes('incoming-call')
-        ) {
-          console.log('Foreground 이벤트 처리', detail.notification?.data);
-          await notifee.cancelNotification(detail.notification.id!);
-          navigationRef.current?.navigate('IncomingCall');
-        }
-      }
-    });
-
-    // Cold start 시 알림 확인
-    (async () => {
-      const initial = await notifee.getInitialNotification();
-      console.log('Cold start 시 알림 확인 data:', initial?.pressAction);
-    })();
-
-    return () => {
-      unsubscribeFg();
-    };
-  }, []);
-
-  useEffect(() => {
-    async function requestPermissions() {
-      if (Platform.OS !== 'android') {
-        return;
-      }
-
-      // 1) Android 13(API 33)+: POST_NOTIFICATIONS 권한 요청
-      if (Platform.Version >= 33) {
-        const granted = await PermissionsAndroid.request(
-          PermissionsAndroid.PERMISSIONS.POST_NOTIFICATIONS,
-        );
-        if (granted !== PermissionsAndroid.RESULTS.GRANTED) {
-          Alert.alert(
-            '알림 권한 필요',
-            '설정 → 앱 → AiRing → 알림에서 권한을 허용해 주세요.',
-            [
-              {text: '취소', style: 'cancel'},
-              {text: '설정으로 이동', onPress: () => Linking.openSettings()},
-            ],
-          );
-        }
-      }
-
-      // 2) Android 12(API 31)+: Exact Alarm 권한 확인 및 안내
-      if (Platform.Version >= 31) {
-        try {
-          const settings = await notifee.getNotificationSettings();
-          if (settings.android?.alarm === AndroidNotificationSetting.ENABLED) {
-            console.log('Exact Alarm 허용됨');
-          } else {
-            Alert.alert(
-              '정확한 알람 권한 필요',
-              '앱이 종료되거나 절전 모드에서도 알람이 울리려면 Exact Alarm을 켜야 합니다.',
-              [
-                {text: '취소', style: 'cancel'},
-                {
-                  text: '설정으로 이동',
-                  onPress: () => notifee.openAlarmPermissionSettings(),
-                },
-              ],
-            );
-          }
-        } catch (e) {
-          console.error('Exact Alarm 권한 확인 중 에러:', e);
-        }
-      }
-    }
-
-    requestPermissions();
-  }, []);
-
-  // 채널 생성
-  useEffect(() => {
-    createVibrationChannels();
-  }, []);
 
   if (isAuthLoading || isAppLockLoading) {
     return <SplashScreen />;
