+++ frontend/src/hooks/useVibrationChannels.ts
@@ -1,0 +1,11 @@
+import {useEffect} from 'react';
+
+import {createVibrationChannels} from '../utils/alarmManager';
+
+const useVibrationChannels = () => {
+  useEffect(() => {
+    createVibrationChannels();
+  }, []);
+};
+
+export default useVibrationChannels;
